requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
dj-static==0.0.6
Django==4.1.5
django-stdimage==6.0.1
gunicorn==20.1.0
Pillow
psycopg2-binary
sqlparse==0.4.3
static3==0.7.0
tzdata==2022.7
